add m3gnetmaker, refactor to move task doc into schema
