+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "fireworks>=2.0.3",
     "atomate>=1.0.3",
     "atomate2>=0.0.8",
+    "m3gnet>=0.1.0",
 ]
 description = "MPmorph is a collection of tools to run and analyze ab-initio molecular dynamics (AIMD) calculations run with VASP, and is currently under development. It relies heavily on tools developed by the Materials Project (pymatgen, custodian, fireworks) and atomate."
 readme = "README.md"
@@ -41,6 +42,7 @@ strict = [
     "fireworks==2.0.3",
     "atomate==1.0.3",
     "atomate2==0.0.8",
+    "m3gnet==0.1.0",
 
 ]
 tests = ["pytest==7.1.3", "pytest-cov==4.0.0"]
