+++ src/mpmorph/workflows/converge.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 from fireworks import Workflow
 from mpmorph.fireworks import powerups
 from mpmorph.fireworks.core import MDFW
-from mpmorph.util import recursive_update
+from mpmorph.utils import recursive_update
 
 __author__ = 'Eric Sivonxay, Jianli Cheng, and Muratahan Aykol'
 __maintainer__ = 'Eric Sivonxay'
